refactor: Perubahan struktur folder dan file
