+++ src/client/src/assets/styles/globals.css
@@ -5,12 +5,13 @@
 }
 
 /* Custom Font */
-@import url("https://fonts.googleapis.com/css2?family=Poppins:ital,wght@0,100;0,200;0,300;0,400;0,500;0,600;0,700;0,800;0,900;1,100;1,200;1,300;1,400;1,500;1,600;1,700;1,800;1,900&display=swap");
-@import url("https://fonts.googleapis.com/css2?family=Poppins:ital,wght@0,100;0,200;0,300;0,400;0,500;0,600;0,700;0,800;0,900;1,100;1,200;1,300;1,400;1,500;1,600;1,700;1,800;1,900&family=Quicksand:wght@300..700&display=swap");
+@import url("https://fonts.googleapis.com/css2?family=Poppins:wght@400;500;600;700&display=swap");
+@import url("https://fonts.googleapis.com/css2?family=Quicksand:wght@400;500;600&display=swap");
+@import url("https://fonts.googleapis.com/css2?family=Rubik+Mono+One&display=swap");
 
 /* Custom Config DaisyUI */
 @utility btn {
-  @apply rounded-lg;
+  @apply rounded-lg bg-primary text-white hover:bg-opacity-90 transition duration-200;
 }
 
 /* Custom Config Tailwindcss */
@@ -20,4 +21,5 @@
 
   --font-poppins: "Poppins", sans-serif;
   --font-quicksand: "Quicksand", sans-serif;
+  --font-rubik: "Rubik Mono One", monospace;
 }
