+++ src/client/src/components/Main/HowItWorks.jsx
